Updated- final

+++ utils.py
@@ -5,8 +5,18 @@ from jinja2 import Template
 
 #Generating new page. 
 def newpage_generator():
-    new_content = """<h1>New Content!</h1>\n<p>New content...</p>"""
-    open("./content/new_content_page.html", "w+").write(new_content)
+    confirmation = input("Do you want to choose your own filename and content to add ? Type 'yes' to confirm and 'no' to use the default: ")
+    if confirmation.lower() == "yes":
+        input_file_name = input("Enter new file name you of your choice with the html extension: ")
+        input_content = input("Enter valid HTML content: ")
+        open("./content/" + input_file_name, "w+").write(input_content)
+        main() # Building site post new page info
+    elif confirmation.lower() == "no":
+        new_content = "<h1>New Content!</h1>\n<p>New content...</p>"
+        open("./content/new_content_page.html", "w+").write(new_content)
+        main() # Building site post new page info
+    else:
+        print("Please choose 'yes' or 'no'.")
 
 
 # Function to check which page to build. Refactored.
@@ -14,19 +24,16 @@ def page_identifier(pages, page_title, page_filename, page_output):
     file_html = open(page_filename).read() 
     template = Template(open("./templates/base.html").read()) 
     results = template.render(
-        pages=pages,                    # added pages list for templating
+        pages=pages,                    # Added pages list for templating
         title=page_title,
-        content=file_html,              #  Try using dictionary
+        content=file_html,              
     )
     open(page_output, 'w+').write(results)
 
 
-# Resume page showing up as first tab even though opened page is index.html
-
+# Generating pages - list of dictionaries
 def list_generate():  
     pages = []
-
-    # Photography title to be replaced by blog. 
     
     all_html_files = glob.glob("content/*.html")
 
@@ -47,7 +54,7 @@ def list_generate():
     return pages
 
 
-
+# main function
 def main():
 
     pages = list_generate()
